Add progress indicator for signup
